# Faster and cleaner exit from hackrf_sweep